fix(#730): break equipment-rows/cards/descriptors circular dependency
LineItemData/GroupData/SubHireGroupData/MixedGroupSlot/CategoryData move to a
new equipment-row-types.ts (pure types, no React/Convex imports).
equipment-row-descriptors.ts and equipment-cards.tsx now import the types from
there instead of from equipment-rows.tsx, which is what closed the 3-cycle loop
(equipment-rows -> {descriptors,cards,line-assets-indicator} -> equipment-rows).
equipment-rows.tsx re-exports the types so existing consumers are unaffected.

POLICY.md R-3.5.
